Finished audit throughput and added tests. Can't test it until the persistence is being loaded into the DI container.

+++ src/Particular.ThroughputCollector.Persistence/EndpointThroughput.cs
@@ -5,6 +5,6 @@
     public class EndpointThroughput
     {
         public DateTime DateUTC { get; set; }
-        public int TotalThroughput { get; set; }
+        public long TotalThroughput { get; set; }
     }
 }
